+++ waiter-app/Comanda/app/src/main/java/dev/giuseppedarro/comanda/features/orders/data/remote/dto/GetOrderResponse.kt
@@ -9,7 +9,7 @@ data class GetOrderResponse(
     val status: String, // "open" | "served" | "closed"
     val items: List<OrderResponseItem>,
     val createdAt: String, // ISO instant timestamp
-    val subtotal: Double? = null,
-    val serviceCharge: Double? = null,
-    val total: Double? = null
+    val subtotal: Int? = null,
+    val serviceCharge: Int? = null,
+    val total: Int? = null
 )
